backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-limiter==3.5.0
pymongo==4.6.1
python-dotenv==1.0.0
eventlet==0.35.1
cryptography==42.0.0
